Composer update

+++ wp-content/plugins/js_composer/include/classes/shortcodes/vc-media-grid.php
@@ -24,7 +24,7 @@ class WPBakeryShortCode_VC_Media_Grid extends WPBakeryShortCode_VC_Basic_Grid {
 			'include' => $atts['include'],
 			'posts_per_page' => apply_filters( 'vc_basic_grid_max_items', self::$default_max_items ),
 			'offset' => 0,
-			'post_type' => 'attachment', // $atts['post_type'],
+			'post_type' => 'attachment',
 			'orderby' => 'post__in',
 		);
 
@@ -40,7 +40,7 @@ class WPBakeryShortCode_VC_Media_Grid extends WPBakeryShortCode_VC_Basic_Grid {
 			$grid_item->shortcodes();
 			$item = '[vc_gitem]<img src="' . vc_asset_url( 'vc/vc_gitem_image.png' ) . '">[/vc_gitem]';
 			$grid_item->parseTemplate( $item );
-			$items = str_repeat( $grid_item->renderItem( get_post( vc_request_param( 'vc_post_id' ) ) ), 3 );
+			$items = str_repeat( $grid_item->renderItem( get_post( (int)vc_request_param( 'vc_post_id' ) ) ), 3 );
 		}
 
 		return $items;
